infrastructure/modules/dataset/python/requirements.txt
google-cloud-logging
google-cloud-bigquery
pyarrow
pandas